FIxes for rescaling and last path segment
When re-scaling the original descriptor enough time the MSEG will get corrupted and shrunk with enough resize updates.. so it's best to keep a separate paint descriptor and virtual MSEG pixel size.
The very last point of the MSEG path also needs to be drawn to as part of the draw-path otherwise in some cases there will be a blank section of the curve (when a point is added close to the end point).